Split canvas-authoring logic into thunderforge_canvas_core for real tests
The engine crate only targets wasm32-unknown-unknown, and this
environment has no wasm-bindgen-test-runner — every engine-side test
written this session only compile-checked, never actually ran.

Add crates/thunderforge-canvas-core: pure data/geometry (Wall, DoorState,
WallEdit, WallSet, is_visible's shadow-casting/segment-intersection
math), zero Bevy/wasm-bindgen dependency, only glam (the identical type
underlying bevy::prelude::Vec2, so values pass between crates with no
conversion). Its 18 wall tests now execute for real via native
`cargo test -p thunderforge_canvas_core`.

Rewrite src/engine/src/resources/wall.rs into a thin Bevy Resource
newtype (Deref/DerefMut) over the core crate's WallSet, so Bevy's
ResMut change detection and every existing call site (wall_set.upsert,
.dirty, etc.) keep working unchanged.

Add ADR-038 recording the decision and establishing the pattern for
lighting/shapes (not yet built) to follow from the start. Updated
plan.md's Project Structure and tasks.md's T036/T052 accordingly.

Verified: cargo test -p thunderforge_canvas_core (18/18 pass, native);
cargo check -p thunderforge_engine --target wasm32-unknown-unknown
(clean, same warning count as before); cargo test -p thunderforge
(41/41 still pass, unaffected).
